✨ Update CI workflow with lint job name and add docker as a dev dependency; remove obsolete test_lock_backends.py
